+++ Professions/Framework/VirtualProperties/Machine_CachedCalibrations.cs
@@ -9,14 +9,14 @@ using System.Runtime.CompilerServices;
 // ReSharper disable once InconsistentNaming
 internal static class Machine_CachedCalibrations
 {
-    internal static ConditionalWeakTable<SObject, Dictionary<string, int>> Values { get; } = [];
+    internal static ConditionalWeakTable<SObject, Dictionary<string, float>> Values { get; } = [];
 
-    internal static Dictionary<string, int> Get_Calibrations(this SObject machine)
+    internal static Dictionary<string, float> Get_Calibrations(this SObject machine)
     {
         return Values.TryGetValue(machine, out var cached) ? cached : [];
     }
 
-    internal static void Set_Calibrations(this SObject machine, Dictionary<string, int> calibrations)
+    internal static void Set_Calibrations(this SObject machine, Dictionary<string, float> calibrations)
     {
         Values.Remove(machine);
         Values.Add(machine, calibrations);
